tap_esc: moved common functions to tap_esc_common
