+++ src/drivers/tap_esc/tap_esc_common.cpp
@@ -36,12 +36,24 @@
  *
  */
 
-#include <px4_config.h>
 #include "tap_esc_common.h"
 
+#include <fcntl.h>
+#include <termios.h>
+
+#include <systemlib/px4_macros.h> // arraySize
+#include <px4_config.h>
+#include <px4_defines.h>
+
+#ifndef B250000
+#define B250000 250000
+#endif
 
 namespace tap_esc_common
 {
+static uint8_t crc8_esc(uint8_t *p, uint8_t len);
+static uint8_t crc_packet(EscPacket &p);
+
 /****************************************************************************
  * Name: select_responder
  *
@@ -58,6 +70,214 @@ void select_responder(uint8_t sel)
 	px4_arch_gpiowrite(GPIO_S2, sel & 4);
 #endif
 }
+
+int initialise_uart(const char *const device, int &uart_fd)
+{
+	// open uart
+	uart_fd = open(device, O_RDWR | O_NOCTTY | O_NONBLOCK);
+	int termios_state = -1;
+
+	if (uart_fd < 0) {
+		PX4_ERR("failed to open uart device!");
+		return -1;
+	}
+
+	// set baud rate
+	int speed = B250000;
+	struct termios uart_config;
+	tcgetattr(uart_fd, &uart_config);
+
+	// clear ONLCR flag (which appends a CR for every LF)
+	uart_config.c_oflag &= ~ONLCR;
+
+	// set baud rate
+	if (cfsetispeed(&uart_config, speed) < 0 || cfsetospeed(&uart_config, speed) < 0) {
+		PX4_ERR("failed to set baudrate for %s: %d\n", device, termios_state);
+		close(uart_fd);
+		return -1;
+	}
+
+	if ((termios_state = tcsetattr(uart_fd, TCSANOW, &uart_config)) < 0) {
+		PX4_ERR("tcsetattr failed for %s\n", device);
+		close(uart_fd);
+		return -1;
+	}
+
+	// setup output flow control
+	if (enable_flow_control(uart_fd, false)) {
+		PX4_WARN("hardware flow disable failed");
+	}
+
+	return uart_fd;
+}
+
+int deinitialise_uart(int &uart_fd)
+{
+	return close(uart_fd);
+}
+
+int enable_flow_control(int uart_fd, bool enabled)
+{
+	struct termios uart_config;
+
+	int ret = tcgetattr(uart_fd, &uart_config);
+
+	if (ret != 0) {
+		PX4_ERR("error getting uart configuration");
+		return ret;
+	}
+
+	if (enabled) {
+		uart_config.c_cflag |= CRTSCTS;
+
+	} else {
+		uart_config.c_cflag &= ~CRTSCTS;
+	}
+
+	return tcsetattr(uart_fd, TCSANOW, &uart_config);
+}
+
+int send_packet(int uart_fd, EscPacket &packet, int responder)
+{
+	if (responder >= 0) {
+		select_responder(responder);
+	}
+
+	int packet_len = crc_packet(packet);
+	int ret = ::write(uart_fd, &packet.head, packet_len);
+
+	if (ret != packet_len) {
+		PX4_WARN("TX ERROR: ret: %d, errno: %d", ret, errno);
+	}
+
+	return ret;
+}
+
+void read_data_from_uart(int uart_fd, ESC_UART_BUF *const uart_buf)
+{
+	uint8_t tmp_serial_buf[UART_BUFFER_SIZE];
+
+	int len =::read(uart_fd, tmp_serial_buf, arraySize(tmp_serial_buf));
+
+	if (len > 0 && (uart_buf->dat_cnt + len < UART_BUFFER_SIZE)) {
+		for (int i = 0; i < len; i++) {
+			uart_buf->esc_feedback_buf[uart_buf->tail++] = tmp_serial_buf[i];
+			uart_buf->dat_cnt++;
+
+			if (uart_buf->tail >= UART_BUFFER_SIZE) {
+				uart_buf->tail = 0;
+			}
+		}
+	}
+}
+
+bool parse_tap_esc_feedback(ESC_UART_BUF *const serial_buf, EscPacket *const packetdata)
+{
+	static PARSR_ESC_STATE state = HEAD;
+	static uint8_t data_index = 0;
+	static uint8_t crc_data_cal;
+
+	if (serial_buf->dat_cnt > 0) {
+		int count = serial_buf->dat_cnt;
+
+		for (int i = 0; i < count; i++) {
+			switch (state) {
+			case HEAD:
+				if (serial_buf->esc_feedback_buf[serial_buf->head] == 0xFE) {
+					packetdata->head = 0xFE; //just_keep the format
+					state = LEN;
+				}
+
+				break;
+
+			case LEN:
+				if (serial_buf->esc_feedback_buf[serial_buf->head] < sizeof(packetdata->d)) {
+					packetdata->len = serial_buf->esc_feedback_buf[serial_buf->head];
+					state = ID;
+
+				} else {
+					state = HEAD;
+				}
+
+				break;
+
+			case ID:
+				if (serial_buf->esc_feedback_buf[serial_buf->head] < ESCBUS_MSG_ID_MAX_NUM) {
+					packetdata->msg_id = serial_buf->esc_feedback_buf[serial_buf->head];
+					data_index = 0;
+					state = DATA;
+
+				} else {
+					state = HEAD;
+				}
+
+				break;
+
+			case DATA:
+				packetdata->d.bytes[data_index++] = serial_buf->esc_feedback_buf[serial_buf->head];
+
+				if (data_index >= packetdata->len) {
+
+					crc_data_cal = crc8_esc((uint8_t *)(&packetdata->len), packetdata->len + 2);
+					state = CRC;
+				}
+
+				break;
+
+			case CRC:
+				if (crc_data_cal == serial_buf->esc_feedback_buf[serial_buf->head]) {
+					packetdata->crc_data = serial_buf->esc_feedback_buf[serial_buf->head];
+
+					if (++serial_buf->head >= UART_BUFFER_SIZE) {
+						serial_buf->head = 0;
+					}
+
+					serial_buf->dat_cnt--;
+					state = HEAD;
+					return true;
+				}
+
+				state = HEAD;
+				break;
+
+			default:
+				state = HEAD;
+				break;
+
+			}
+
+			if (++serial_buf->head >= UART_BUFFER_SIZE) {
+				serial_buf->head = 0;
+			}
+
+			serial_buf->dat_cnt--;
+
+		}
+
+
+	}
+
+	return false;
+}
+
+static uint8_t crc8_esc(uint8_t *p, uint8_t len)
+{
+	uint8_t crc = 0;
+
+	for (uint8_t i = 0; i < len; i++) {
+		crc = crc_table[crc^*p++];
+	}
+
+	return crc;
+}
+
+static uint8_t crc_packet(EscPacket &p)
+{
+	/* Calculate the crc over Len,ID,data */
+	p.d.bytes[p.len] = crc8_esc(&p.len, p.len + 2);
+	return p.len + offsetof(EscPacket, d) + 1;
+}
+
 
 const uint8_t crc_table[256] = {
 	0x00, 0xE7, 0x29, 0xCE, 0x52, 0xB5, 0x7B, 0x9C, 0xA4, 0x43, 0x8D, 0x6A,
